fix: change the response varibale type of the id

+++ internal/features/prescription/prescription.service.go
@@ -10,7 +10,7 @@ func NewService(repo Repository) *Service {
 	return &Service{repo}
 }
 
-func (this *Service) Create(pres *Prescription) (int64, error) {
+func (this *Service) Create(pres *Prescription) (int, error) {
 	newID, err := this.repo.Create(pres)
 	if err != nil {
 		return 0, fmt.Errorf("error al crear prescription en repo: %w", err)
